requirements.txt
asgiref==3.4.1
black==21.9b0
click==8.0.3
Django==3.2.8
flake8==4.0.1
isort==5.9.3
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.19.2
pytz==2021.3
regex==2021.10.8
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
